Don't crash when serializing empty CSSUnparsedValues
The "tokens_" list (which despite its name is really a list of
CSSUnparsedSegments) can be non-empty, but still produce an empty
sequence when tokenized. We're not taking this into account, and can
therefore end up with an empty CSSParserTokenRange, which will DCHECK
when passed to CSSVariableData::Create.

To fix this, move the empty-check until after the CSSParserTokenRange
has been created.

Fixed: 1169941
Change-Id: I3eac2b46dca9999ba5f4b509f8293f604f678198